requirements.txt
streamlit
streamlit-feedback
pandas
python-dotenv
openai
